+++ nilearn-stubs/maskers/multi_nifti_maps_masker.pyi
@@ -17,7 +17,7 @@ class MultiNiftiMapsMasker:
         detrend: int | bool = ...,
         low_pass: float | None = ...,
         high_pass: float | None = ...,
-        t_r: int | None = ...,
+        t_r: float | None = ...,
         dtype: int | None = ...,
         resampling_target: int | str | None = ...,
         memory: Memory | str | Path | None = ...,
