+++ nilearn-stubs/maskers/nifti_maps_masker.pyi
@@ -20,7 +20,7 @@ class NiftiMapsMasker:
         detrend: int | bool = ...,
         low_pass: float | None = ...,
         high_pass: float | None = ...,
-        t_r: int | None = ...,
+        t_r: float | None = ...,
         dtype: int | None = ...,
         resampling_target: int | str | None = ...,
         keep_masked_maps: int | bool = ...,
